host_frame: gs_host_frame_object_get_locked(): introduce and make use of it
This function will be re-used in a later commit.
